pyproject.toml
[project]
name = "graphrag-customer-journey"
version = "0.1.0"
description = "Graph-RAG for Customer Journey Intelligence - E-commerce product analytics with graph-enhanced retrieval"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "networkx>=3.2",
    "pandas>=2.1",
    "numpy>=1.26",
    "faker>=22.0",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "groq>=0.4",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "streamlit>=1.30",
    "python-dotenv>=1.0",
    "httpx>=0.26",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=24.1",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
